requirements.txt
pandas>=2.0
rich>=13.0
click>=8.0
prompt_toolkit>=3.0
PyYAML>=6.0
openpyxl>=3.1
chardet>=5.0
jinja2>=3.1
